04-DEC

+++ demo/src/Test.java
@@ -2,7 +2,7 @@
 public class Test {
 
 	public static void main(String[] args) {
-		float f = 10.50;
+		double f =  10.50;
 
 	}
